+++ util/messages.h
@@ -1,6 +1,9 @@
 /*
  * Prototypes for message and error reporting (possibly fatal).
  *
+ * The canonical version of this file is maintained in the rra-c-util package,
+ * which can be found at <http://www.eyrie.org/~eagle/software/rra-c-util/>.
+ *
  * Copyright 2008, 2010
  *     The Board of Trustees of the Leland Stanford Junior University
  * Copyright (c) 2004, 2005, 2006
@@ -91,7 +94,7 @@ void message_log_syslog_crit(size_t, const char *, va_list, int)
     __attribute__((__nonnull__));
 
 /* The type of a message handler. */
-typedef void (*message_handler_func)(size_t, const char *, va_list, int);
+typedef void (*message_handler_func)(unsigned int, const char *, va_list, int);
 
 /* If non-NULL, called before exit and its return value passed to exit. */
 extern int (*message_fatal_cleanup)(void);
